Fix classId type mismatch in accounting + add server setup doc
Build fix:
- Class.id is Integer (INT in schema), but new accounting code declared
  classId as Long throughout — caused compile errors:
    FeePricingService.java:43 incompatible types Integer→Long
    InvoiceService.java:222 incompatible types Integer→Long
    + cascading "cannot find getClassName() on Object" from broken
      stream type inference
- Aligned classId to Integer in:
    entity/FeePricing.java
    dto/FeePricingRow.java
    dto/InvoiceGenerationRequest.java
    repository/FeePricingRepository.java
    repository/EnrollmentRepository.java
    service/FeePricingService.java (Map<Integer, BigDecimal>,
                                    Map<Integer, FeePricing>)
    service/InvoiceService.java (List<Integer>, dummy -1 literal,
                                 nested Map<Long, Map<Integer, BigDecimal>>,
                                 local classId var)
- Schema migration appended at bottom of sql/accounts_schema.sql:
  drop FK → MODIFY class_id INT → re-add FK. Idempotent, works on
  fresh deploys and on existing BIGINT deployments.

Docs:
- New docs/server-setup.md — end-to-end school server bring-up:
  prereqs, MySQL setup + technonix grant, secret generation
  (JWT_SECRET, PLATFORM_METRICS_SECRET, DB_PASSWORD), .env creation,
  schema run order, run options (Docker / Maven), verification curls,
  production checklist, troubleshooting.

+++ src/main/java/com/example/academic_service/dto/InvoiceGenerationRequest.java
@@ -20,5 +20,5 @@ public class InvoiceGenerationRequest {
     private Integer academicYearId;
 
     /** Optional class ids filter. Null/empty = all classes. */
-    private List<Long> classIds;
+    private List<Integer> classIds;
 }
